requirements.txt
pandas
numpy
scikit-learn
joblib
flask
flask-wtf
werkzeug
python-dotenv
